refactor: Moving lts api controllers
